requirements.txt
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
Scrapy==2.12.0